git-svn-id: svn://192.168.1.10/public/CoffeeMud@8827 0d6f1817-ed0e-0410-87c9-987e46238f29

+++ com/planet_ink/coffee_mud/WebMacros/BankAccountInfo.java
@@ -92,7 +92,7 @@ public class BankAccountInfo extends StdWebMacro
 				playerM.setLocation(M.location());
 				playerM.setStartRoom(M.getStartRoom());
 				playerM.setClanID(C.clanID());
-				playerM.setClanRole(C.getTopRankedRoles(Function.DEPOSIT_LIST).get(0));
+				playerM.setClanRole(C.getTopRankedRoles(Function.DEPOSIT_LIST).get(0).intValue());
 				destroyPlayer=true;
 			}
 			else
